Stage 1: barycentric_tex RG16float -> RGBA16uint
First step of the GPU-instancing pipeline plan. Repacks the barycentric
attachment to hold both barycentric.xy (now u16 fixed-point in RG
channels) and a future per-fragment instance_id (BA channels, split u32
via join32). Stays at 4 color attachments — no risk of exceeding the
WebGPU `maxColorAttachments = 4` lower bound.

Stage-1 is the format change only. The geometry vertex shader emits
U32_MAX as a placeholder instance_id; the fragment shader packs that
into BA; the compute pass reads it but doesn't use it yet. Subsequent
stages wire the real instance_id flow + per-instance attribute lookup.

Changes:
- RenderTextureFormats.barycentric: Rg16float -> Rgba16uint
- material_opaque bind_group.rs: sample_type UnfilterableFloat -> Uint
- bind_groups.wgsl (opaque): texture_multisampled_2d<f32> -> <u32>
- geometry fragment.wgsl: FragmentOutput.barycentric vec2<f32> -> vec4<u32>;
  pack bary as `u32(bary.xy * 65535.0 + 0.5)` in RG; instance_id (sentinel
  for now) packed via split16 into BA
- geometry vertex.wgsl: VertexOutput gains `instance_id: u32` (flat),
  emits U32_MAX for now
- compute.wgsl + material_shading.wgsl MSAA path: unpack RG as
  `vec2<f32>(f32(raw.x), f32(raw.y)) / 65535.0`; bind raw _instance_id
  via join32 for Stage-2 consumer

u16 fixed-point gives uniform ~1.5e-5 precision in [0, 1] — comparable
to f16 over the same range.

Verified: arena renders identically (all primitives, particles,
sweep, lines, instances, sprites, toon, orbit camera all unchanged).

+++ crates/renderer/src/render_passes/material_opaque/shader/material_opaque_wgsl/bind_groups.wgsl
@@ -1,12 +1,13 @@
 {% if multisampled_geometry %}
     @group(0) @binding(0) var visibility_data_tex: texture_multisampled_2d<u32>;
-    @group(0) @binding(1) var barycentric_tex: texture_multisampled_2d<f32>;
+    // Barycentric tex packs: RG = bary.xy as u16 fixed-point, BA = instance_id (split u32).
+    @group(0) @binding(1) var barycentric_tex: texture_multisampled_2d<u32>;
     @group(0) @binding(2) var depth_tex: texture_depth_multisampled_2d;
     @group(0) @binding(3) var normal_tangent_tex: texture_multisampled_2d<f32>;
     @group(0) @binding(4) var barycentric_derivatives_tex: texture_multisampled_2d<f32>;
 {% else %}
     @group(0) @binding(0) var visibility_data_tex: texture_2d<u32>;
-    @group(0) @binding(1) var barycentric_tex: texture_2d<f32>;
+    @group(0) @binding(1) var barycentric_tex: texture_2d<u32>;
     @group(0) @binding(2) var depth_tex: texture_depth_2d;
     @group(0) @binding(3) var normal_tangent_tex: texture_2d<f32>;
     @group(0) @binding(4) var barycentric_derivatives_tex: texture_2d<f32>;
